Update of symlinks script generator

+++ generate_symlinks.sh
@@ -10,6 +10,14 @@ fi
 path=$1
 
 for link in {"data","results","saved_models","models_infos","models_backup","threshold_map","learned_zones","custom_norm"}; do
-    rm ${link}
-    ln -s $1/${link} ${link}
+    
+    if [ -L ${link} ]; then
+        rm ${link}
+    fi
+    
+    if [ ! -d "${path}/${link}" ]; then
+        mkdir -p ${path}/${link}
+    fi
+    
+    ln -s ${path}/${link} ${link}
 done
